fix(di): limit circular dependencies
