+++ src/main/java/org/radarbase/appserver/service/scheduler/DataMessageSchedulerService.java
@@ -21,13 +21,9 @@
 
 package org.radarbase.appserver.service.scheduler;
 
-import com.google.firebase.messaging.FirebaseMessagingException;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
 import org.radarbase.appserver.entity.DataMessage;
-import org.radarbase.appserver.service.FcmDataMessageService;
-import org.radarbase.appserver.service.FcmNotificationService;
-import org.radarbase.appserver.service.UserService;
 import org.radarbase.appserver.service.scheduler.quartz.SchedulerService;
 import org.radarbase.fcm.downstream.FcmSender;
 import org.radarbase.fcm.model.FcmDataMessage;
@@ -49,11 +45,8 @@ public class DataMessageSchedulerService extends MessageSchedulerService<DataMes
 
     public DataMessageSchedulerService(
             @Autowired @Qualifier("fcmSenderProps") FcmSender fcmSender,
-            @Autowired SchedulerService schedulerService,
-            @Autowired UserService userService,
-            @Autowired FcmNotificationService notificationService,
-            @Autowired FcmDataMessageService dataMessageService) {
-        super(fcmSender, schedulerService, userService, notificationService, dataMessageService);
+            @Autowired SchedulerService schedulerService) {
+        super(fcmSender, schedulerService);
     }
 
     private static FcmDataMessage createMessageFromDataMessage(DataMessage dataMessage) {
@@ -74,13 +67,6 @@ public class DataMessageSchedulerService extends MessageSchedulerService<DataMes
     }
 
     public void send(DataMessage dataMessage) throws Exception {
-        try {
-            fcmSender.send(createMessageFromDataMessage(dataMessage));
-        } catch (FirebaseMessagingException exc) {
-            log.error("Error occurred when sending downstream message.", exc);
-            // TODO: update the data message status using event
-            handleErrorCode(exc.getErrorCode(), dataMessage);
-            handleFCMErrorCode(exc.getMessagingErrorCode(), dataMessage);
-        }
+        fcmSender.send(createMessageFromDataMessage(dataMessage));
     }
 }
