+++ src/test/java/org/radarbase/appserver/service/scheduler/NotificationSchedulerServiceTest.java
@@ -52,10 +52,7 @@ import org.radarbase.appserver.config.SchedulerConfig;
 import org.radarbase.appserver.entity.Notification;
 import org.radarbase.appserver.entity.Project;
 import org.radarbase.appserver.entity.User;
-import org.radarbase.appserver.service.FcmDataMessageService;
-import org.radarbase.appserver.service.FcmNotificationService;
 import org.radarbase.appserver.service.MessageType;
-import org.radarbase.appserver.service.UserService;
 import org.radarbase.appserver.service.scheduler.NotificationSchedulerServiceTest.SchedulerServiceTestConfig;
 import org.radarbase.appserver.service.scheduler.quartz.SchedulerServiceImpl;
 import org.radarbase.fcm.downstream.FcmSender;
@@ -223,9 +220,7 @@ class NotificationSchedulerServiceTest {
 
             // mock FCM as we do not want to connect to the server
             return new NotificationSchedulerService(
-                    mock(FcmSender.class), new SchedulerServiceImpl(scheduler),
-                    mock(UserService.class), mock(FcmNotificationService.class),
-                    mock(FcmDataMessageService.class));
+                    mock(FcmSender.class), new SchedulerServiceImpl(scheduler));
         }
     }
